first fri re-sort
