+++ client_sdk/fri/include/fri/friClientApplication.h
