+++ client_sdk/fri/include/fri/friConnectionIf.h
