+++ client_sdk/fri/package.xml
@@ -3,8 +3,9 @@
 <package format="3">
   <name>fri</name>
   <version>1.0.0</version>
-  <description>KUKA fast robot interface with ROS2 build system</description>
+  <description>KUKA fast robot interface with ROS 2 build system.</description>
   <maintainer email="martin.huber@kcl.ac.uk">mhubii</maintainer>
+  <maintainer email="christopher.mower@kcl.ac.uk">cmower</maintainer>
   <license>see LICENSE</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
