+++ client_sdk/fri/src/FRIMessages.pb.c
