--- client_sdk/fri/src/client_lbr/Makefile
@@ -1,13 +1,0 @@
-BASE_DIR	= ../..
-include $(BASE_DIR)/build/GNUMake/paths.mak
-include $(BASE_DIR)/build/GNUMake/$(TOOLS_MAK)
-
-CXX_SRC 		 = friLBRClient.cpp \
-			 		friLBRCommand.cpp \
-			 		friLBRState.cpp
-
-INC_DIR		+= $(CLIENTBASE_DIR) $(NANOPB_DIR) $(PROTOBUF_DIR) $(PROTOBUF_GEN_DIR)
-CXXFLAGS 	+= 
-LDFLAGS 	+= 
-
-include $(BASE_DIR)/build/GNUMake/rules.mak
