--- client_sdk/fri/src/connection/Makefile
@@ -1,14 +1,0 @@
-BASE_DIR	= ../..
-include $(BASE_DIR)/build/GNUMake/paths.mak
-include $(BASE_DIR)/build/GNUMake/$(TOOLS_MAK)
-
-CXX_SRC 		 = friUdpConnection.cpp
-
-CXXFLAGS 	+= 
-LDFLAGS 		+= 
-
-################################################################################
-### Include general makefile (at the end)
-################################################################################
-
-include $(BASE_DIR)/build/GNUMake/rules.mak
