+++ client_sdk/fri/src/friCommandMessageEncoder.cpp
@@ -5,10 +5,10 @@ agreement or other license is obtained by KUKA Deutschland GmbH, Augsburg, Germa
 
 SCOPE
 
-The software “KUKA Sunrise.FRI Client SDK” is targeted to work in
-conjunction with the “KUKA Sunrise.FRI” toolkit.
-In the following, the term “software” refers to all material directly
-belonging to the provided SDK “Software development kit”, particularly source
+The software ï¿½KUKA Sunrise.FRI Client SDKï¿½ is targeted to work in
+conjunction with the ï¿½KUKA Sunrise.FRIï¿½ toolkit.
+In the following, the term ï¿½softwareï¿½ refers to all material directly
+belonging to the provided SDK ï¿½Software development kitï¿½, particularly source
 code, libraries, binaries, manuals and technical documentation.
 
 COPYRIGHT
@@ -58,85 +58,69 @@ cost of any service and repair.
 \version {2.5}
 */
 #include <cstdio>
-#include "friMonitoringMessageDecoder.h"
-#include "pb_decode.h"
 
+#include "friCommandMessageEncoder.h"
+#include "nanopb/pb_encode.h"
 
 using namespace KUKA::FRI;
 
 //******************************************************************************
-MonitoringMessageDecoder::MonitoringMessageDecoder(FRIMonitoringMessage* pMessage, int num)
+CommandMessageEncoder::CommandMessageEncoder(FRICommandMessage* pMessage, int num)
    : m_nNum(num), m_pMessage(pMessage)
 {
    initMessage();
 }
 
 //******************************************************************************
-MonitoringMessageDecoder::~MonitoringMessageDecoder()
+CommandMessageEncoder::~CommandMessageEncoder()
 {
 
 }
 
 //******************************************************************************
-void MonitoringMessageDecoder::initMessage()
+void CommandMessageEncoder::initMessage()
 {
-   // set initial data
-   // it is assumed that no robot information and monitoring data is available and therefore the 
-   // optional fields are initialized with false
-   m_pMessage->has_robotInfo = false;
-   m_pMessage->has_monitorData = false;
-   m_pMessage->has_connectionInfo = true;
-   m_pMessage->has_ipoData = false;
-   m_pMessage->requestedTransformations_count = 0;
+   m_pMessage->has_commandData = false;
    m_pMessage->has_endOfMessageData = false;
-   
-   
+   m_pMessage->commandData.has_jointPosition = false;
+   m_pMessage->commandData.has_cartesianWrenchFeedForward = false;
+   m_pMessage->commandData.has_jointTorque = false;
+   m_pMessage->commandData.has_cartesianPose = false;
+   m_pMessage->commandData.commandedTransformations_count = 0;
+   m_pMessage->commandData.has_redundancyInformation = false;
    m_pMessage->header.messageIdentifier = 0;
-   m_pMessage->header.reflectedSequenceCounter = 0;
+   // init with 0. Necessary for creating the correct reflected sequence count in the monitoring msg
    m_pMessage->header.sequenceCounter = 0;
+   m_pMessage->header.reflectedSequenceCounter = 0;
 
-   m_pMessage->connectionInfo.sessionState = FRISessionState_IDLE;
-   m_pMessage->connectionInfo.quality = FRIConnectionQuality_POOR;
-
-   m_pMessage->monitorData.readIORequest_count = 0;
-   m_pMessage->monitorData.measuredCartesianPose.element_count = 0;
+   m_pMessage->commandData.writeIORequest_count = 0;
 
    // allocate and map memory for protobuf repeated structures
-   map_repeatedDouble(FRI_MANAGER_NANOPB_DECODE, m_nNum, 
-         &m_pMessage->monitorData.measuredJointPosition.value,
-         &m_tSendContainer.m_AxQMsrLocal);
+   map_repeatedDouble(FRI_MANAGER_NANOPB_ENCODE, m_nNum, 
+         &m_pMessage->commandData.jointPosition.value,
+         &m_tRecvContainer.jointPosition);
+   map_repeatedDouble(FRI_MANAGER_NANOPB_ENCODE, m_nNum, 
+         &m_pMessage->commandData.jointTorque.value,
+         &m_tRecvContainer.jointTorque);
    
-   map_repeatedDouble(FRI_MANAGER_NANOPB_DECODE, m_nNum, 
-         &m_pMessage->monitorData.measuredTorque.value,
-         &m_tSendContainer.m_AxTauMsrLocal);
-   
-   map_repeatedDouble(FRI_MANAGER_NANOPB_DECODE, m_nNum, 
-         &m_pMessage->monitorData.commandedTorque.value,
-         &m_tSendContainer.m_AxTauCmdLocal);
-   
-   map_repeatedDouble(FRI_MANAGER_NANOPB_DECODE, m_nNum, 
-         &m_pMessage->monitorData.externalTorque.value,
-         &m_tSendContainer.m_AxTauExtMsrLocal);
-   
-   map_repeatedDouble(FRI_MANAGER_NANOPB_DECODE,m_nNum,
-         &m_pMessage->ipoData.jointPosition.value,
-         &m_tSendContainer.m_AxQCmdIPO);
-
-   map_repeatedInt(FRI_MANAGER_NANOPB_DECODE, m_nNum, 
-         &m_pMessage->robotInfo.driveState,
-         &m_tSendContainer.m_AxDriveStateLocal);
+   // nanopb encoding needs to know how many elements the static array contains
+   // a quaternion always contains 7 elements
+   m_pMessage->commandData.cartesianPose.element_count = 7;
+   // a Cartesian wrench feed forward vector always contains 6 elements
+   m_pMessage->commandData.cartesianWrenchFeedForward.element_count = 6;
 }
 
 //******************************************************************************
-bool MonitoringMessageDecoder::decode(char* buffer, int size)
+bool CommandMessageEncoder::encode(char* buffer, int& size)
 {
-    pb_istream_t stream = pb_istream_from_buffer((uint8_t*)buffer, size);
-
-    bool status = pb_decode(&stream, FRIMonitoringMessage_fields, m_pMessage);
+    // generate stream for encoding
+    pb_ostream_t stream = pb_ostream_from_buffer((uint8_t*)buffer, FRI_COMMAND_MSG_MAX_SIZE);
+    // encode monitoring Message to stream
+    bool status = pb_encode(&stream, FRICommandMessage_fields, m_pMessage);
+    size = stream.bytes_written;
     if (!status)
     {
-        printf("!!decoding error on Monitor message: %s!!\n", PB_GET_ERROR(&stream));
+        printf("!!encoding error on Command message: %s!!\n", PB_GET_ERROR(&stream));
     }
-
     return status;
 }
