+++ client_sdk/fri/src/friCommandMessageEncoder.h
