+++ client_sdk/fri/src/friLBRCommand.cpp
@@ -5,10 +5,10 @@ agreement or other license is obtained by KUKA Deutschland GmbH, Augsburg, Germa
 
 SCOPE
 
-The software “KUKA Sunrise.FRI Client SDK” is targeted to work in
-conjunction with the “KUKA Sunrise.FRI” toolkit.
-In the following, the term “software” refers to all material directly
-belonging to the provided SDK “Software development kit”, particularly source
+The software ï¿½KUKA Sunrise.FRI Client SDKï¿½ is targeted to work in
+conjunction with the ï¿½KUKA Sunrise.FRIï¿½ toolkit.
+In the following, the term ï¿½softwareï¿½ refers to all material directly
+belonging to the provided SDK ï¿½Software development kitï¿½, particularly source
 code, libraries, binaries, manuals and technical documentation.
 
 COPYRIGHT
@@ -57,11 +57,12 @@ cost of any service and repair.
 \file
 \version {2.5}
 */
-#include "friLBRState.h"
-#include "friLBRCommand.h"
+#include "fri/friLBRState.h"
+#include "fri/friLBRCommand.h"
+#include "fri/friDataHelper.h"
+
 #include "friClientData.h"
 #include "pb_frimessages_callbacks.h"
-#include "friDataHelper.h"
 
 using namespace KUKA::FRI;
