+++ client_sdk/fri/src/friMonitoringMessageDecoder.h
