+++ client_sdk/fri/src/friUdpConnection.cpp
@@ -5,10 +5,10 @@ agreement or other license is obtained by KUKA Deutschland GmbH, Augsburg, Germa
 
 SCOPE
 
-The software “KUKA Sunrise.FRI Client SDK” is targeted to work in
-conjunction with the “KUKA Sunrise.FRI” toolkit.
-In the following, the term “software” refers to all material directly
-belonging to the provided SDK “Software development kit”, particularly source
+The software "KUKA Sunrise.FRI Client SDK" is targeted to work in
+conjunction with the "KUKA Sunrise.FRI" toolkit.
+In the following, the term "software" refers to all material directly
+belonging to the provided SDK "Software development kit", particularly source
 code, libraries, binaries, manuals and technical documentation.
 
 COPYRIGHT
@@ -63,7 +63,7 @@ cost of any service and repair.
 #include <unistd.h>
 #endif
 
-#include "friUdpConnection.h"
+#include "fri/friUdpConnection.h"
 
 
 #ifdef WIN32
