+++ client_sdk/fri/src/pb_frimessages_callbacks.c
@@ -5,10 +5,10 @@ agreement or other license is obtained by KUKA Deutschland GmbH, Augsburg, Germa
 
 SCOPE
 
-The software “KUKA Sunrise.FRI Client SDK” is targeted to work in
-conjunction with the “KUKA Sunrise.FRI” toolkit.
-In the following, the term “software” refers to all material directly
-belonging to the provided SDK “Software development kit”, particularly source
+The software ï¿½KUKA Sunrise.FRI Client SDKï¿½ is targeted to work in
+conjunction with the ï¿½KUKA Sunrise.FRIï¿½ toolkit.
+In the following, the term ï¿½softwareï¿½ refers to all material directly
+belonging to the provided SDK ï¿½Software development kitï¿½, particularly source
 code, libraries, binaries, manuals and technical documentation.
 
 COPYRIGHT
@@ -61,9 +61,10 @@ cost of any service and repair.
 #include <stdio.h>
 #include <stdlib.h> 
 
+#include "nanopb/pb_encode.h"
+#include "nanopb/pb_decode.h"
+
 #include "pb_frimessages_callbacks.h"
-#include "pb_encode.h"
-#include "pb_decode.h"
 
 bool encode_repeatedDouble(pb_ostream_t *stream, const pb_field_t *field, void * const *arg)
 {
