--- client_sdk/fri/src/protobuf_gen/Makefile
@@ -1,15 +1,0 @@
-BASE_DIR	= ../..
-include $(BASE_DIR)/build/GNUMake/paths.mak
-include $(BASE_DIR)/build/GNUMake/$(TOOLS_MAK)
-
-CC_SRC 		 = FRIMessages.pb.c
-
-INC_DIR		+= $(NANOPB_DIR)
-CFLAGS 		+= 
-LDFLAGS 		+= 
-
-################################################################################
-### Include general makefile (at the end)
-################################################################################
-
-include $(BASE_DIR)/build/GNUMake/rules.mak
